Fix unmounting and signal handling on OS X
On OS X a FUSE volume needs to be unmounted before exiting the event loop.
Unmounting the volume flushes the kernel's unified buffer cache, which results
in file system requests, that need to be handled by the event loop.

Exiting the event loop without unmounting causes the file system daemon to hang
for daemon_timeout seconds (defaults to 60 seconds) and can result in data loss
or corruption.