Add support for huge (4 Petabyte) max cache size
Existing riofs caches only support up to 4 GBytes of cached
data, as cache_dir_max_size is a 32 bit unsigned integer
count of the maximum cache size, in bytes.

Extend that by a factor of a million (1024 * 1024, actually).

Existing riofs configurations still work, as is, with no
change in allowed cache size.  A new configuration option
can optionally be used, to override the existing setting,
with a count of how many MBytes, instead of how many bytes,
are allowed in the max cache directory size.

Remove "filesystem.cache_dir_max_size" from the list of
config settings that must be set on riofs startup, because
now either that setting, and/or the new setting of
"filesystem.cache_dir_max_megabyte_size" may be set
(with the latter one taking precedence, if set.)

The code that determines the configured size will complain
if neither option is set.