ETag fix #3: add ETag consistency checking
This is the one very important patch in this series.

The proper way, in my understanding, to check whether
the file we have locally cached is still the same file
as on the AWS S3 server is to compare ETag's.

Anytime we notice that our cached file no longer has
the same exact contents as the similarly named file on
the AWS S3 server, we need to invalidate our cache and
download the new file contents as need be.

AWS S3 guarantees (to the level of confidence of MD5 sums)
that if and only if the content of a stored file changes,
then its ETag changes.  There is no need to compute MD5
sums locally (which aren't the same as ETags on multipart
files anyway) to make this check.  If the ETag hasn't
changed, we're still dealing with the same file contents,
even if we have only cached a few parts of it so far.

Other ways of checking, such as on file size or version or
calculated MD5 sums for single part files, are quite less
reliable, subject to both false positives and false negatives,
and not always applicable, if for example a file is not
versioned on Amazon, or a file has multiple parts, or a
file's contents changes but its size remained the same.

This patch makes the following changes:
 1) Add an etag string to CacheEntry, to track what ETag
    is currently in the cache for that inode.
 2) Add get and update routines, to access that etag string.
 3) Add an aws_etag string to FileReadData, to keep track
    of what was reported in the latest ETag header field
    from the AWS S3 server for that file, so that once the
    corresponding CacheEntry is available as well, we can
    get, update, and compare the two etags, and invalidate
    the cache if necessary.
 4) Add insure_cache_etag_consistent_or_invalidate_cache()
    routine, called from two places, depending on the
    sequencing of events, to handle this ETag checking.
 5) The insure_cache_etag_consistent_or_invalidate_cache()
    routine returns TRUE if it was able to make the check
    and (if necessary), invalidate the cache.  It returns
    FALSE if and only if it encountered an error along the way.
 6) The previously unused evkeyvalq argument to the
    fileio_read_on_get_cb() routine is now used, as
    the ETag header is needed in that routine.
 7) FileReadData *rdata can no longer be free'd using a
    single g_free() call, as it might contain an allocated
    aws_etag.  So add a fileread_destroy() routine to
    handle free'ing it.
 8) Add a cache_etag_is_set flag to FileReadData, to keep
    track of whether we've set the etag string in the
    associated (via the ino number) CacheEntry.  On newly
    cached files, setting up the cache occurs some steps
    after setting up FileReadData, so we need to keep track
    of whether we've set the cache's etag or not.
 9) A few other minor changes in related comments and spacing.

This patch provides a potentially substantial performance
improvement for a particular use case.  If a new large file
was placed on an riofs managed server, and multiple clients
learn of it and all go to download it at nearly the same
(downloads overlap in time), then dreadful cache thrashing
could occur, as the cache kept being invalidated when another
client came along to start a download, and the size of
whatever had been stored in the cache so far did not match
the total size of what was on the AWS S3 server.  Repeated
partial downloads of the same file could result in S3
bandwidth charges that were one or two orders of magnitude
larger than needed for a single download of such a file.