ETag fix #2: split off fileio size setting
Some key code in the fileio_read_on_head_cb() routine
was serving a dual purpose:
 1) Set the file's size from the http header Content-Length
 2) Perform one of the consistency checks, this one for size change

After adding the ETag consistency checking code in the next patch,
I will then be able to remove the other, less reliable, methods
of checking, including checking on file size.   However the
setting of the file's size from the http header is still needed.

Therefore in this patch I split up this code, into two separate
parts, one of which will remain, and the other of which will
be removed in a few patches.